commit 23:54 27/06

+++ gestaoDePatrimonio/src/view/EditPatrimony.java
@@ -419,7 +419,7 @@ public class EditPatrimony implements ActionListener
 					{
 						try
 						{
-							area = Integer.parseInt(textFields.get(4).getText());
+							area = Double.parseDouble(textFields.get(4).getText());
 							
 							verVH[4] = true;
